In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,9 +10,9 @@
 int main(void)
 {
 	int i;
-	unsigned int fib1 = 1, fib2 = 2, nextFib;
+	unsigned long int fib1 = 1, fib2 = 2, nextFib;
 
-	printf("%u, %u, ", fib1, fib2);
+	printf("%lu, %lu, ", fib1, fib2);
 
 	for (i = 2; i < 98; i++)
 	{
@@ -22,11 +22,11 @@ int main(void)
 
 		if (i < 97)
 		{
-			printf("%u, ", nextFib);
+			printf("%lu, ", nextFib);
 		}
 		else
 		{
-			printf("%u\n", nextFib);
+			printf("%lu\n", nextFib);
 		}
 	}
